% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.decum.R
\name{fun.decum}
\alias{fun.decum}
\title{Decumulate Values in a Matrix}
\usage{
fun.decum(matrix)
}
\arguments{
\item{matrix}{This is a matrix of values to be decumulated where the cumulation occurs over rows within the same column.}
}
\description{
The function takes a matrix of values that have accumulated over rows and
returns a matrix of the incremental increase between each row. Do not run fun.decum
as a stand-alone function
}
\examples{
testarray <- array(NA, dim=c(5,2,2))
testarray[,,1] <- cbind(c(1:5), c(1:5)*2)
testarray[,,2] <- cbind(c(1:5)*1.5, c(1:5)*3)
fun.decum(testarray)
}
\keyword{sequential}
