% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.exposure.R
\name{fun.exposure}
\alias{fun.exposure}
\title{Define Exposure Accumulation Function}
\usage{
fun.exposure(rate, offset, t)
}
\arguments{
\item{rate}{Rate of accumulation.}

\item{offset}{Initial exposed population.}

\item{t}{Time at which individuals are exposed.}
}
\description{
This function creates a dataset simulating the accumulation of individuals exposed to 
treatment under the self-controlled risk interval design.
}
\examples{
fun.exposure(rate=100, offset=20, t=20)
}
\keyword{exposure}
\keyword{sequential}
\keyword{simulation}
