% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCRI.seq.r
\name{SCRI.seq}
\alias{SCRI.seq}
\title{Perform sequential analysis on true and misclassified binomial data}
\usage{
SCRI.seq(data, params)
}
\arguments{
\item{data}{Output from sim.exposure that contains real and observed data.}

\item{params}{Output from initialize.data function.}
}
\description{
This function performs a prespecified number of binomial sequential analyses on real and misclassified binomial data that is designed to simulate a self-controlled risk interval design. Do not run SCRI.seq
as a stand-alone function.
}
\examples{
#paramtest <- initialize.data(seed=8768, N=1, t0=0, tf=2, NStrata=2, 
#strataRatio=c(0.2, 0.3, 0.3, 0.2), EventRate=c(0.4, 0.5), sensitivity=0.9, PPVest=0.9, RR=3.0, 
#MatchRatio=1, maxSampleSize=200, maxTest=1, totalAlpha=0.05, minEvents=5, AlphaSpendType="Wald",
#AlphaParameter=0.5, address=getwd(), rate=20, offset=30)
#exposed1 <- create.exposure(paramtest)
#exposed2 <- sim.exposure(exposed1, paramtest)
#SCRI.seq(exposed2, paramtest)
}
\keyword{sequential}
